+++ app/Http/Controllers/TaskController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Task;
+use App\Project;
 
 class TaskController extends Controller
 {
@@ -41,6 +42,9 @@ class TaskController extends Controller
             'due_by' => 'required'
         ]);
         
+        $projectId = $request->input('project_id');
+        $this->authorize('interact', Project::find($projectId));
+        
         Task::create(request()->all());
 
         return response()->json('ok', 200);
@@ -77,6 +81,8 @@ class TaskController extends Controller
      */
     public function update(Request $request, Task $task)
     {
+        $this->authorize('interact', Project::find($task->project_id));
+
         $task->update($request->all());
 
         return response()->json('ok', 200);
@@ -92,9 +98,11 @@ class TaskController extends Controller
     {
         if ($request->query('deleteAllCompletedTasksOfAssociatedProject')) {
             $projectId = $request->query('projectId');
+            $this->authorize('interact', Project::find($projectId));
             $ids = Task::where('project_id', $projectId)->where('completed', true)->pluck('id');
             Task::destroy($ids);
         } else {
+            $this->authorize('interact', Task::find($id));
             Task::destroy($id);
         }
